Corrected app structure.
