+++ src/app/MainApp.java
@@ -3,6 +3,7 @@ package app;
 import app.view.NavigationMenuController;
 import app.view.WindowsNavigator;
 import javafx.application.Application;
+import javafx.collections.ObservableList;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.image.Image;
