+++ src/app/model/Person.java
@@ -1,0 +1,13 @@
+package app.model;
+
+public class Person extends CommonDirectory {
+
+    public Person(String name, boolean defaultPerson) {
+        super(null, name, defaultPerson);
+    }
+
+    public Person(Integer id, String name, boolean defaultPerson) {
+        super(id, name, defaultPerson);
+    }
+
+}
