+++ src/app/repository/PersonRepository.java
@@ -1,0 +1,13 @@
+package app.repository;
+
+import app.model.Person;
+
+import java.util.List;
+
+public interface PersonRepository {
+    void save(Person person);
+
+    void delete(int id);
+
+    List<Person> getAll();
+}
