+++ src/app/repository/SiteRepository.java
@@ -1,0 +1,13 @@
+package app.repository;
+
+import app.model.Site;
+
+import java.util.List;
+
+public interface SiteRepository {
+    void save(Site site);
+
+    void delete(int id);
+
+    List<Site> getAll();
+}
