+++ src/app/service/KeywordService.java
@@ -1,0 +1,12 @@
+package app.service;
+
+import app.model.Keyword;
+import javafx.collections.ObservableList;
+
+public interface KeywordService {
+    void save(Keyword keyword);
+
+    void delete(int id);
+
+    ObservableList<Keyword> getAllByPersonId(Integer id);
+}
