--- src/main/java/repository/SiteRepository.java
@@ -1,13 +1,0 @@
-package main.java.repository;
-
-import main.java.model.Site;
-
-import java.util.List;
-
-public interface SiteRepository {
-    void save(Site site);
-
-    void delete(int id);
-
-    List<Site> getAll();
-}
