--- src/main/java/service/KeywordService.java
@@ -1,12 +1,0 @@
-package main.java.service;
-
-import javafx.collections.ObservableList;
-import main.java.model.Keyword;
-
-public interface KeywordService {
-    void save(Keyword keyword);
-
-    void delete(int id);
-
-    ObservableList<Keyword> getAllByPersonId(Integer id);
-}
